pantalla lista pedido jaja

+++ src/Pedido.java
@@ -17,11 +17,11 @@ public class Pedido {
     private int diaCreacion;
     private int mesCreacion;
     private int anioCreacion;
-    private int demoraEstimada;
+    private long demoraEstimada;
     private ItemPedido itemPedido;
     private final int ANIO_CREACION = 2018;
 
-    public Pedido(String nombreCliente, int diaCreacion, int mesCreacion, int anioCreacion, int demoraEstimada, ItemPedido itemPedido) throws Exception {
+    public Pedido(String nombreCliente, int diaCreacion, int mesCreacion, int anioCreacion, long demoraEstimada, ItemPedido itemPedido) throws Exception {
         if (nombreCliente.isEmpty() || nombreCliente == null || nombreCliente.contains("0") && nombreCliente.contains("1") && nombreCliente.contains("2") && nombreCliente.contains("3") && nombreCliente.contains("4") && nombreCliente.contains("5") && nombreCliente.contains("6") && nombreCliente.contains("7") && nombreCliente.contains("8") && nombreCliente.contains("9")) {
             throw new Exception("Ingrese un nombre de cliente válido");
         }
